+++ fastestimator/dataset/__init__.py
@@ -15,3 +15,4 @@
 from fastestimator.dataset.generator_dataset import GeneratorDataset
 from fastestimator.dataset.labeled_dir_dataset import LabeledDirDataset, LabeledDirDatasets
 from fastestimator.dataset.numpy_dataset import NumpyDataset
+from fastestimator.dataset.op_dataset import OpDataset
